% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharaoh.R
\name{create_pyramid}
\alias{create_pyramid}
\title{create_pyramid}
\usage{
create_pyramid(dimension, initial_value = 0, step = 0)
}
\arguments{
\item{dimension}{Matrix dimension}

\item{initial_value}{Matrix initial value}

\item{step}{Height delta between levels of pyramid}
}
\description{
Creates pyramid height matrix
}
